+++ src/main/java/com/github/danielfabela/partpicker/menus/Menu.java
@@ -6,38 +6,27 @@ public class Menu{
 
     // Method to display the main Menu
     public void MainMenu(){
-        System.out.println("Welcome to Part Picker");
-        System.out.println("*************************************");
+        System.out.println("*************************************************************");
+        System.out.println("Welcome to PC Budget");
+        System.out.println("*************************************************************");
         System.out.println("Main Menu");
-        System.out.println("Choose your parts wisely");
         System.out.println("0. Quit");
-        System.out.println("1. CPU");
-        System.out.println("2. MotherBoard");
-        System.out.println("3. GPU");
-        System.out.println("4. RAM");
-        System.out.println("5. SSD");
-        System.out.println("6. HDD");
-        System.out.println("7. Power Supply");
-        System.out.println("8. Case");
-        System.out.println("*************************************");
-        System.out.println("Input your choice. Only numbers from 0 to 11.");
+        System.out.println("1. See Listed Components");
+        System.out.println("2. Calculate Total Price of Each Part");
+        System.out.println("3. Display Total Price for PC");
+        System.out.println("*************************************************************");
+        System.out.println("Input your choice. Only numbers from 0 to 3.");
     }
 
-    public void CPUMenu(){
-        System.out.println("*************************************");
-            System.out.println("Choose your CPU");
-            System.out.println("1. Model 1");
-            System.out.println("2. Model 2");
-            System.out.println("*************************************");
-            System.out.println("Input your choice. Only numbers from 0 to 2.");
+    public void ComponentsMenu(){
+        System.out.println("*************************************************************");
+            System.out.println("Items in your list");
+            System.out.println("    Part  |                Model              |  Price  ");
+           
     }
 
-    public void motherboardMenu(){
-        System.out.println("*************************************");
-            System.out.println("Choose your MotherBoard");
-            System.out.println("1. Model 1");
-            System.out.println("2. Model 2");
-            System.out.println("*************************************");
-            System.out.println("Input your choice. Only numbers from 0 to 2.");
+    public void ChooseMenu(){
+        System.out.println("*************************************************************");
+        System.out.println("Input the Id's of the products you want to purchase");
     }
 }