+++ src/main/resources/schema.sql
@@ -1,0 +1,8 @@
+create table components (
+    id serial primary key,
+    Part varchar,
+    Model varchar,
+    Price varchar,
+    Taxes decimal,
+    Total decimal
+)